requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
flask-migrate==4.0.4
flask-jwt-extended==4.4.4
python-dotenv==1.0.0
mysql-connector-python==8.0.33
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
python-dateutil==2.8.2